fix: extract to a fn for test
